tools/CodeNaturalnessEvaluator/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.23.0=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.12.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
autoconf=2.71=pl5340h5eee18b_0
autoconf-archive=2021.02.19=ha770c72_0
bitsandbytes=0.41.1=pypi_0
bitsandbytes-cuda117=0.26.0.post2=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cpuonly=2.0=0
cryptography=41.0.3=py310hdda0065_0
datasets=2.14.5=pypi_0
dill=0.3.7=pypi_0
evaluate=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=py310h06a4308_0
install=1.3.5=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py310h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kenlm=0.2.0=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
m4=1.4.18=h516909a_1001
markupsafe=2.1.1=py310h7f8727e_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
nltk=3.8.1=pypi_0
numpy=1.25.2=py310h5f9d8c6_0
numpy-base=1.25.2=py310hb5e798b_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.10=h7f8727e_2
packaging=23.1=pypi_0
pandas=2.1.0=pypi_0
perl=5.32.1=2_h7f98852_perl5
pillow=9.4.0=py310h6a678d5_0
pip=23.2.1=py310h06a4308_0
psutil=5.9.5=pypi_0
pyarrow=13.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=2.0.1=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.8.8=pypi_0
requests=2.31.0=py310h06a4308_0
responses=0.18.0=pypi_0
safetensors=0.3.3=pypi_0
scipy=1.11.2=pypi_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.0=pypi_0
torchaudio=2.0.2=py310_cpu
torchvision=0.15.2=py310_cpu
tqdm=4.66.1=pypi_0
transformers=4.34.0.dev0=pypi_0
tree-sitter=0.20.2=pypi_0
typing_extensions=4.7.1=py310h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=py310h06a4308_0
wheel=0.38.4=py310h06a4308_0
xxhash=3.3.0=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
